# Conflicts:
#	Garamgaebi/app/src/main/java/com/garamgaebi/garamgaebi/common/XAccessTokenInterceptor.kt
#	Garamgaebi/app/src/main/java/com/garamgaebi/garamgaebi/src/main/register/LoginActivity.kt

+++ Garamgaebi/app/src/main/java/com/garamgaebi/garamgaebi/model/LoginResponse.kt
@@ -12,7 +12,7 @@ data class LoginResult(
     val grantType: String,
     val memberIdx: Int,
     val refreshToken: String,
-    val refreshTokenExpirationTime: Int
+    val refreshTokenExpirationTime: Long
 )
 
 data class LoginRequest(
